refactor: share DS formatting
